Move method to InstructorsLogic and add test
