+++ src/test/java/teammates/ui/webapi/UpdateInstructorActionTest.java
@@ -45,7 +45,7 @@ public class UpdateInstructorActionTest extends BaseActionTest<UpdateInstructorA
 
         String newInstructorName = "newName";
         String newInstructorEmail = "newEmail@email.com";
-        String newInstructorRole = "Tutor";
+        String newInstructorRole = Const.InstructorPermissionRoleNames.INSTRUCTOR_PERMISSION_ROLE_TUTOR;
 
         InstructorCreateRequest reqBody = new InstructorCreateRequest(instructorId, newInstructorName,
                 newInstructorEmail, newInstructorRole,
